Completed 2-3.
Did not follow Hint at end of lesson (not sure how that would help.)

My implementation uses a lot of wildcards `_` so all but `b_not/1` accept any time of value and will return false if given a non-boolean value.  I am not sure if this is a good idea.

I could write write a `is_bool/1` function which would return true if the arguments is the atoms `true` or `false`.