Cleaned up boolean module.
Removed error return Tuple from b_not/1.  It fails if the passed value is not `true` or `false`.

Removed the "short-circuit" from b_and/2.  It was a bad idea.

Rewrote b_nand/2.  I had misunderstood the logic (thought it was exclusive or.)   Now written using the hint from the book.