Completed 3-7.
Rewrote the db module using built-in lists module functions.

new/0, destroy/1, and write/3 are all unchanged.  The body of delete/2 went down to one line and no helper functions.  read/2 was reduced to a single case statement with two cases.

I could not immediately figure out how to remove the helper function match/3 for match/2.  I suspect that I could have accomplished this using lists:map or something similar, but as it is, it is still greately reduced.