Finished 4-2: Process Ring in first try!
This exercise was easier than I thought it would be.  I chose to use the centralized implementation (strategy 1).

ring:start/3 uses two helper functions to create the series of linked nodes and then to send the message.

Writing ring:createNode/1 this way simplified the overall development of the module, I believe.

ring:sendMessage/3 only knows how to talk to its node parameter and handles converting start/3's Message parameter into a tuple that loop/1 knows what to do with.  Finally, if the number of message iterations falls to 0, it intializes the node destruction process by telling the FirstNode to stop.

ring:loop/1 handles printing the message and passing it on to its Pid parameter.

Closes #2.