Complete 5-1 my_db module.
This exercise was also fairly easy to complete following the two hints in the book.  I had a brief hiccup in loop/1 when I forgot that it needed to pass the result of calling the appropriate db module function back to the process that messaged it (instead of just returning the result).  This was fairly simple to solve by adding Pid to the message protocol and receive clauses to the message passing functions.

I think I'm starting to get the hang of message passing and persisting state in the receive/respond loop.

Closes #15.